requirements.txt
requests
beautifulsoup4
pandas
openai
anthropic
python-dotenv
rich
flask
flask-cors
selenium
webdriver-manager
